Remove support for LLVM 13
This is in preparation for LLVM support. There are many functions in
LLVM 13 that are deprecated in LLVM 14 and removed in LLVM 15.
Replacements are only in LLVM 14, so it would be difficult to support
both LLVM 13 and LLVM 15. Removing LLVM 13 support is the first step
towards incrementally adding LLVM 15 support.